Tweak totals calculation to handle records with NULL trxn_id.
After running the scripts there is a small number (< 50k est) of contacts who are still not
showing totals. For most of these it is because they have NULL trxn_id transactions which
are being missed.

Deployment of trigger update = no hurry. Can deploy & then ops can load when ready

Bug: T186165
Change-Id: Ifffcf323d1a79624223b44423152c558b0d16be6