Move hack to block search without criteria from core-hack to extension.
This is still pretty unlovable but I don't think it will be upstreamed any time soon so better
in an extension than as a core hack. I don't know when or if it is ever hit these days
(we could throw an exception instead if we wanted to answer that).

Bug: T194210
Change-Id: I57373bac1a0ec9452442c460155f683779fb4097