Fix to ensure that setting source to empty does not cause a hard-error.
The anonymise extension sets source to 'null' which is handled as NULL in the DAO. If we then
try to update contribution extra we get a hard error.

We might decide not to let anonymise set source to 'null'/ NULL but it seems reasonable to
not try to update wmf_contribution_extra if we don't have a valid update.

Change-Id: Ib8f0ebb374b76d4a02611cc7e5f4304037bded3f