Move metadata to class, switch to whitelisting of entities to show & forget.
The metadata controls what entities & fields we show & forget. I had been hoping for more of a
'delete everything except these few things' but it seems there is quite a lot of internal 'noise'
and I felt I had to go through each entity with contact data & look at whether to show it, to
delete it or to ignore it. The metadata includes all entities that FK onto contact id,
whether or not they seem relevant. However, I commented out all those that did not seem relevant.

I think this makes the implicit decisions fairly explicit - hopefully.

If I'm honest I still don't know if this class will be the 'final home' of the metadata - it
feels like something that should be in a json or yaml format - however, if it moves again
the class & functions to retrieve it can just wrap that retrieval.

I wanted to move it to make it more accessible to the logging.showme & forgetme
as I kinda need to iterate through entities on that too.

Bug: T196644
Change-Id: I352183cc31233cdeb1a6ce7d39c44a20870ebe12