Fix processor_id for subscriptions not yet canceled
This is the easy part - these people have signed up within the past
month (+3 days) and their contribution_recur status is still either
Completed (all good) or Failed (had a problem in the last couple days).

For these people, we can just reset the processor_id and let the
recurring_globalcollect job take it from there. For the canceled
ones, we'll want a bit more logic to include the RecurringRestarted
tag in their next thank you email.

As of 2018-07-19 there are 691 Completed and 19 Failed WebCollect
recurring txns with bad processor_id.

Bug: T199331
Change-Id: I43711447812891d682c9d1bd5bc3464b8f948e17