Add RecurringRestartedUncharged tag
This one is for the civicrm_contribution_recur table. It will
indicate that the subscription has been cured of some malady,
and that the next contribution record created from it should
get the RecurringRestarted tag so the donor is thanked correctly.
The recurring processor should then remove this tag from the
civicrm_contribution_recur table.

Bug: T199331
Change-Id: I06e14d1f84377924836f09986c23d91702bef2af