Fix gateway reconciliation failing when is_negative group by in play.
Defining the field alias allows it to be used in the group by clause.
I'm unsure why this would have regressed - unless the report criteria
change to expose an existing regression.

Note there is a bit of a 'weakness' that 'fields' metadata doubles as
general metadata so this needs to be definited in fields. In extended reports &
to a lesser extent in core I've tried to move to an actual metadata key.

I'm going to try to do a follow up cleanup patch but wanted this to stand
alone

Bug: T202319
Change-Id: I21ceb2e5d653f6703f6ddb43e737afe18a4590b1