Add no_thank_you default for new contribs via UI
The new contribution form is fetched in two http requests. The
first calls the buildForm hook with CRM_Contribute_Form_Contribution
and has the main contribution fields and the soft credit fields.

The second calls the buildForm hook with CRM_Custom_Form_CustomDataByType,
and has the custom fields such as wmf_contribution_extra (where
no_thank_you lives). The second buildForm hook doesn't have an
_action set like the first one does, so we have to check for the
presence of an _entityId to tell whether we're editing an existing
contribution or adding a new one.

Bug: T202415
Change-Id: If5e3683ea81ceb1b90409f3ba0af7d59540f5efb