Move _extra field calc to pre hook
To avoid inserting twice. This way they're inserted along with the
source_ fields that are passed into the contribution creation api
call.

Because we're potentially adding custom fields in the middle of an
api call, we not only add the custom_XX keys to the contribution
params, we also have to format them as is done at the start of
civicrm_api3_contribution_create with _civicrm_api3_custom_format_params,
which copies the values and a bunch of metadata to the 'custom'
key.

Bug: T203537
Change-Id: Iced4a3e05b6c5289f7137fe395da10cce763cbbd