Exclude anonymous contact from triggers.
This causes an early return on any totals calculation triggers for insert & update for
- Anonymous contact (72 on live)
- Financial Type Endowment gift

I left off the delete & alter-old contact triggers for now as this is primarily for performance
in this iteration.

I included the first part of excluding Endowment Gift contributions, as much for ease as
anything. I can't be sure dev installs will have an anonymous donor but
the Endowment Gift financial type is created in wmf_civicrm.install. So by
including the Endowment Gift I
1) create some code signposts while my head in in that code
2) save myself writing a bunch of conditional code for when there
are no required conditions

Bug: T199753
Change-Id: Id6a1710387520df0f2bdfc985f4274bc656f81d6