Further upstream fixes for contact layout editor.
For our purposes this fixes
1) tags have lost their colour (this is from a comment not a separate issue, to test you need
to configure a tag with a colour and also have tags in your contact layour block).
2) blocks that don't work on organizations still show up on organization records but
do not work. This is an issue that hurt Elliot & I in testing but is not
hit in live since Nora has configured correctly in live.

I've closed this follow up against the parent issue - the sub issues have been
separately resolved

Bug: T203266
Change-Id: Ie214890bb78281290334233f05070cba0a2f2f01