Switch variables to being class properties
The thinking here is that by making them class properties we can break down the functions
without passing as many variables around.

There are some obvious follow ones in the closeFilesAndSetMessage function signature
and I left them off this round purely to keep the commit size small & readable.

If we can break it up a bit we can aim to switch to https://hardcoredev.com/blog/using-batch-api-in-drupal-7-tutorial-in-simple-code/

Bug T204008

Change-Id: Ic56dcef7bd8dfc056f84673028c44ba2d6c1ce4c