Adjust omnirecipient repair to reflect a daylight savings change.
This re-applies an earlier patch but with the Sent filter now removed

The below is from the earlier commit message
----------------------------------------------------------------------------------------------------
From early Oct 2017 the 7 hour gap stops catching the duplicates & an 8 hour gap starts catching them.

This adjusts the script to the 8 hour gap (will check for any other variant on daylight savings
once that has run through. Note I found this reference row

contact_identifier | mailing_identifier|email|action|date|contact_id
174777435979|sp55350275|###########@gmail.com|Sent|11/14/2017 23:23:48|12840722

In the csv we can see that actual GMT time is 23:23:48.

In the database we have the times of
2017-11-14 02:23:48
2017-11-14 10:23:48

The script deletes the former & fixes the latter - 23 -10 = 13 - so I am adding on 11 hours to
fix those rows where the time gap indicates our altered daylight savings gap

Bug: T200240
Change-Id: I397913a6f84d105315c818afb99bdd1675c74d4b