Delete child activities before parent activities to avoid error on delete.
We have had reports of some contacts that won't delete. On retesting they do. I believe the issue
is merge activities have child activities. When the parent is deleted so is the child,
later the child can't be deleted because it's already gone.

This fix would not work if there is a chain of activity parentage but I'm not aware of
any instances where that is in play. Perhaps in CiviCase it could happen but we don't use
CiviCase

Bug: T204063
Change-Id: I0f5f2fa86dad407e0fa1d8bf11f46d3f77b567bd