Functional import using batch api.
The batch api prevents gateway timeouts by doing multiple requests & has a progress bar.

Notes
1) I exposed the batch limit with a default of 100 - I have no idea what a good number is but MG can
experiment & for dev testing we can set at a low number
2) For testing I focussed on Benevity & Engage (& the tests) - there is a file with an error streak
> 10 called engage_multiple_errors in the test data folder which was useful for testing that
through the UI since the unit test for it is less useful now (as the messaging for the error streak has
been moved to the batch handled)
3)  When using the batch api the batch iterations run in separate php processes so variables in
memory are not available between runs. The basic importer object can be serialised, but not if there
are file resources open on it (they also don't stay open between iterations) so I am sticking with
passing the basic instantiaed object & using the batch api $context mechanism for state data.

I'm not sure if state data would be better as an array than as properties - probably not enough
better to warrant the re-work on that.
4) I did think while doing this that rather than creating output csv it might make more sense to have
a DB table and only generate the output csv on request from that table. We'd need a drush job to
intermittently clear it out.
5) The refund job is the only one that overrides the import function. It doesn't seem to be exposed in the UI

Bug: T204008

Change-Id: Ifa3d99cbd7ebf8b952fd8834ba5ec90e00f88554