Switch to phpleage csv Writer package.
This allows to do transforms when loading a row which is nice & also to get a row
count which we is important for batch processing.

I am also remove the code to close the files. The phpleague package doesn't mention
doing that & seems to handle it iself here possibly -
https://github.com/thephpleague/csv/blob/5e69545d55d440e2862b90e9bf9ba6c91333d41e/src/Stream.php#L154

If not closing files were to be an issue we would see it in tests rather than live so let zuul be my judge.

Bug: T204008
Change-Id: I76f45d76e2469bd5111723649875843693210230