Make output files batch friendly.
The way the importer object is passed between batch iterations is by serialising & unserialising
the object so it breaks if any properties cannot be serialized. Batch runs are separate page requests
so the php memory is flushed between them - only the parameters managed by the batch api hold state
information & these are presumably stored in a table in serialized format between batches.

Since the resources cannot be serialised we output the headers to each file
but rely on re-opening them next batch run rather than passing them.

Note that the passed instance of the importer is the original copy - not one that has been updated
through batch runs.

Bug: T204008

Change-Id: I20b8549515c75f31df7204576b1602e741e3f8f1