Change Omnimail Forgetme so that it writes Erase requests to the database & processes in a separate job
We hit a case where it seems Silverpop did not process the Erase request. This moves the request to the database
where it is retained until it has succeeded.

Note
- a separate cron job needs to call civicrm_api3('Omnirecipient', 'process_forgetme', ['mail_provider' => 'Silverpop']);
- at this stage I have still to fix it so re-tries are given job criteria and only try to fetch if a job already exists
- I'm not sure if a job could stop being fetchable from Silverpop and not be completed - in which case it wouldn't go
from the table, or if we need to handle the situation that occurred when the email was recreated by our export
after being forgotten. If that were the case we'd need to retry again a few days later I guess.

Bug: T199747

Change-Id: I8c2ac759c004d6cc621118808ec4ee7d84667eaa