Add ability to exclude one or more financial types form large notification emails.
This adds a field to the table for 'excluded financial types' and to the notification forms.

It then checks against them when sending out the mails.

I also opted to load the thresholds & filter against them in php rather than a query per new contribution.

Also, on the form - despite not setting it to 'required' I can't seem to deselect all options
from the new field. I decided to pretend I hadn't notices that as I didn't think we would be doing
that any time soon.

Bug: T207712
Change-Id: I413e1da9fbe1c4220b326e2c9370dbbdd2cf765e