Fix indiscriminate showing (and forgetting) of payment tokens.
It turns out the id filter is being dropped :-( due to the DAO having id defined
with a uniqueName. Fix is generic to any fields that use Unique Name.

I was able to replicate this in a test. My expectation is that without this
the payment tokens showm in 'showme' would have been forgotten in 'forgetme'
had a contact with a payment token been forgotten.

However, I'm pretty confident mbeat33 picked this up on the first potential instance
as the payment token id 1 still exists

Bug: T212705
Change-Id: I90f2c8aa13d2b6e77d7278f15b171b7a7958a148