Handling for multiple erase databases
The key change here is
https://github.com/eileenmcnaughton/omnimail-silverpop/commit/d1811a86f5bf0a8ff57164e3ccce020ff2aff1b6

which permits database_id to be passed as an array OR a string.

To use multiple databases either pass in an array or change
civicrm.settings so

$civicrm_setting['Omnimail Preferences']['omnimail_credentials']['Silverpop'] is an array of DB ids

The rest of the change is really just test handling, in part because database_id can no longer
be empty in test scenarios.

The return output is different when an array is passed - which is a bit ugly.
We don't really handle the output that much at all currently so it's
a bit irrelevant in a way & I'm a bit undecided as to whether to put more
time into thinking about / improving this as it's not really affecting variables
that are used

Bug: T205332
Change-Id: I2314271b5323b5013fa7a36d861f06a804e6745e