Update Omnirecipient to UPDATE IGNORE so exact matches don't cause the script to fail
We have one contact (239378476336) whose email client is obviously polling heavily and has over 4000 open records
for just a couple of received emails. The script is trying to clean up a record on this person but with so many
he is hitting an exact timestamp match on the record we are trying to update.

The timestamp is part of the unique index on the table - ergo it fails.

Duplicate open records are not valuable data & tbh any given 20000 records in this table
are of marginal value since it's used for aggreagate stats so we lose nothing here

Change-Id: I20ae7c6211fc7141c372a65001d5813ef3ffa35f