Restore colour coding on dedupe screen
Per https://github.com/civicrm/org.civicrm.shoreditch/issues/373 civi-core colour codes the rows to make
it easier to see difference visually but shoreditch whomps this - perhaps by accident or perhaps the designers
rated look-consistency over usability. Either this restores it in a 'fairly maintainable' way.

This wasn't specifically requested by DS but it ALSO makes it a lot easier to see if related
dev work to re-order the fields is working so I did it first

Bug: T217903
Change-Id: Id8d33f40f06aefe038b64c7f5575a19a66d7e72c