Update Contact Layout Editor.
We have a dependency situation here
- apiv4 ships the latest version with the release
- the latest version of api v4 required an update to CLE
- we can update apiv4 independently of the civi release & doing so allows us to
get these synced prior to the civi code update but we need
to co-ordinate these 2 patches (one in the submodule)
- if there is a problem this fails hard - ie. you get a fatal error
when you go to view a contact. That is really the only thing
we need to be checking/ the only place this will impact us & it's
not subtle.

Bug: T218754
Change-Id: Iadc01ae23afc9f8b9de8330b0d43b338f28457d1