Add cancel_reason to cancel
This adds cancel_reason in 3 places

SmashPig - '(auto) maximum failures reached' when we cancel from repeated retries
RecurringQueueConsumer - '(auto) User Cancelled via Gateway' when Paypal notifies of us
  a txn_type of 'subscr_cancel'
                      -  '(auto) Expiration notification' when we stop trying due to a subscr_eot message (not sure
                         the process by which this happens,

There were no pre-existing tests on the cancel action & a change in core means we have possibly not been
processing paypal cancels since the update (we were using a BAO function not an api function & the sig changed)

This fixes that & adds a test so we are OK in future. I don't know if we need to rerun anything - I doubt we missed much

Note importSubscriptionCancel is only reached when the txn type is subscr_cancel

Potential scenarios. We distill down these trxn_types
- recurring_payment_profile_cancel => subscr_cancel (handled in this commit)
- recurring_payment_expired => subscr_eot - we treat this as the natural end of our subscription - now handled in this commit

- recurring_payment_failed, recurring_payment_skipped, recurring_payment_suspended, recurring_payment_suspended_due_to_max_failed_payment
=> subscr_failed - our watchdog message says we cancel these - I don't think we actually do though from the code. We do in SmashPig but
not RecurringQueueConsumer. This might warrant another Phab? I dont want to go down a rabbit hole on this but once we are doing
better reporting we might see this come up again.

Bug: T218616

Change-Id: I65507834a469020f2c8d775d163f177c88fb22cf