Shrink log check window to a month before audit date
The audit processor is currently taking all the logs from the oldest
transaction in any processed audit file, to the newest transaction,
and searching for all of the missing transactions in all of the log
files before their posted audit date. This patch only searches for
transactions in the log files up to a month before their posted
audit date.

Because iDEAL donors sometimes re-use reference numbers from previous
years, donations coming in this year can be counted by Ingenico as
an additional payment on a donation from years ago, and have an
audit date far far in the past. This has meant the 'earliest'
transaction date was super far back, and left the audit processor
grinding along looking for all of today's donations in multiple
years worth of logs.

Change-Id: Ie3da33d5efc62b831bdc69bc8cf0e69b00efde1b