Use new contribution tracking queue
Anywhere we write to the contribution table (outside of
the queue consumer) is changed to send messages to the
contribution tracking queue. To replace inserts we first
get a new id from the contribution-tracking sequence
generator.

Needs drush updb.

Adds test setup to initialize the sequence generator to
the max ID in the contribution_tracking table, and adds
calls to consume the queue before testing any expectations
about what's in that table.

Bug: T215463
Change-Id: Ib01216db910ee37d2179c1460a5209a03826513a