Remove call to retrieve cachedMergeInfo when updating the url
This call checks if the merge the merge criteria being typed result in merge results & renders them.

It was something I added before the getDuplicates but and it's pretty clear from me from watching
staff use it that it's no longer helpful & just adds confusion.

(I might look at removing more calls to it but this seems clearcut)

T217903

Change-Id: I377739cab4b04fc6519b5002fee48ed286df2b95