build: vary AMPHOME for Docker containers
The legacy Jenkins jobs had to get AMPHOME to point to the build
workspace in order for it to be deleted between builds. We also needed
to vary the path to ensure consecutive builds would not unintentionally
reusing a previous build configuration.

With the future Docker containers, the logic is no more needed. We can
just point AMPHOME to XDG_CONFIG_HOME which is empty at the start of the
container and entirely deleted when the container is stopped.

Keep the old logic around for back compatibility. That will be dropped
once the legacy CI jobs are deleted.
Print out the value of AMPHOME.

Bug: T210287
Change-Id: If464e133eb267ff07ca1bf960623d96ad2ef0a7f