Move creation of wmf_donor fields to update_custom_fields
We are about to add more fields but this change does not do that. It does change the order and
we are not tinkering with the weight to get the UI order right but I feel like that's of no consequence.

Nora does know how to re-order fields in the UI

Bug: T170972
Change-Id: Ic18f2eb04475b92e6133f3fc1a92a289befd5c3d