Get rid of history on do_not_solicit, simplify
I initially was going to remove these functions altogether but then just removed from the install
functions.

On one hand it feels like it makes sense to clean them up. On the other, if we are going to do that maybe
we should agree what to remove first & do it in a more structured way.

Change-Id: Ifeecc28af4f90eb3d01067a1ae25d994d73851a2