Separate do_not_solicit in merge code from otherwise calculated fields
We moved the do_not_solicit out of wmf_donor & this just logically separates it from them in the merge code.

The unit test that covers this is MergeTest::testMergeHook

Note I was trying to dig into why I perceived we needed special handling for some fields during merge
over and above the triggers and that the 2 joined fields - first contribution amount and
date of largest contribution might need special handling because we added some for last_donation_amount, last_donation_usd.

In my digging I updated the test to check these fields on merge and they 'just worked'. I'm wondering whether the
special handling on merge for 'latest' fields is redundant now.

I note we DO need special handling on import for 'last_original_amount' because this is calculated when the contribution
is saved but the custom field is not saved until after than - meaning we have to do a post hook

So, the good news is there might not be anything more to see here....

Bug T222958

Change-Id: I0d49c7daa5b81bf67ad8c9b79ff350337760eed2