Exclude Refund contributions from calculations.
I noticed that we have a non-zero lifetime total for contacts with only a refund -eg.

They pay $3
The request a refund - which winds up being $3.15
The contribution of $3 is updated to 'Refunded' so it is excluded from calculations but the
balance transaction of .15 IS included.

This is trivially more correct - ie. the lifetime total includes the washup on the refund but
it is not meaningfully correct and it also throws out things like 'most recent donation'

I've been struggling to explain the hang we got on triggers based on my testing & I wondered if this could play in
but TBH I can't quite see how getting a different value for largest_donation to join against would
matter.

Also note that the trxn_id could be financial_type_id instead - I didn't switch as it would need testing for performance
- I *think* mysql handles  financial_type_id IN (endowmentType, refundType) as well as = endowmentType
but I recently got some surprising results on BETWEEN vs <= and >=

Bug: T229273
Change-Id: I4e61323eb48203ffd246b4a0d44cfa1642bc2b6e