Add targetsmart extension
This is intended to be used for the batch script but at the moment it is just a shell with
an upgrader function that creates the field mappings.

My intent is to do some tests through the UI (first on staging & then on live) with the mapping &
see how it looks & get MG to do some looking too & then start of the batch side.

Bonus - they will get a handful imported to nosy at in the meantime.

Bug: T228712

Change-Id: Ibeeeee8e8c40f0e365ee8786406629ed2a61db6b