Add api & test for target smart import
Note this adds 2 fields to the field mapping and one group that I manually added on prod while testing. Dev
envs will need some manual attention to this.

This is based on 2 classes I have put up for core - the MetadataTrait is merged and
the ImportProcessor is partially merged
https://github.com/civicrm/civicrm-core/pull/15072
(I'm adding tests & cleanup to core based on it as the 'justification' for it being in core
- in fact the import code is needlessly hard & could really be cleaned up but review enthusiasm is
a little low so will not push too far into it).

Per the test this can be called by

drush --user=1  cvapi TargetSmart.import csv=../sites/default/files/import1.tsv  batch_size=1000 offset=5 debug=1

Note the user needs to be set to avoid an unfortunate permission gotcha

The file that ships in the test works for prod / staging - the contact ids match test contacts
but for local testing the contact id needs editing.

Bug : T228714

Change-Id: Ic8e36c095ab8132a8fad7ece7d1ce20bc6b558cb