Add  extension to manage our query modifications
For years we carried a hack to track the user behind any run-away queries. An alternative
(suggested by Tim) that I put up a long time back has finally been merged
and this utilises that approach - allowing us to drop another patch.  Our  other
patch in the packages directory was merged for 5.18 so this is the last time
we should need to patch  packages

https://github.com/civicrm/civicrm-packages/pull/180

I did some performance tests & while I found the starting point (~460 per minute) seemed slower than when I last did
some tests I got a little better with the 5.17 upgrade (~484 per minute) & no discernable difference when
switching to this approach. Note that the 460 per minute could reflect 'something' going on so there is not
necessarily an improvement with the 5.17 update but not worse is also good.

Bug: T228826

Change-Id: Ib9074b982dd43c706dd08bf577e81bd72335024f