Fix batched import to cope with csv reader upgrade.
This got missed because it's so 'drupally' it didn't have the test cover. I think there is an issue
with the serialization of the  class so I'm reloading the reader each time now.

Other than that the improved header handling meant all the offsets got reduced by one.

I used the engage.csv in this modules test dir for testing

Bug: T231399
Change-Id: Ia31290f627fb95c3e0291e20db687ec6e3800da7