Omnimail test cleanup
I'm having  trouble running the Omnimail tests locally for 'some  reason'.

I  can't quite figure it out but it seems that the trouble is only tests using the
End2End Interface (which emulates a full drupal site) not the Headless interface.

This doesn't change that - it just makes it so the interfaces are only assigned in
one place & removes empty overrides. Also some switches to use api helper functiosn

Bug: T230509
Change-Id: Idbbf70896c48fe00fb98777fca4aa26570d9b8a7