Adapt to change in upstream Processor/ imporrt code
So the tragedy of  the  core import mappings is they store the label not the name field. Import is
really old  code  which is why & hopefully one  day I'll build off the cleanup I've done to fix this.

In the mean time the core code  was  missing some labels & they got added, invalidating some mappings.

For our pusposes this means  Contact ID presenting differently in the values array. We use that to
add to group & for logging

 (the actual import code now expects the label not the name so it is fine but once I added the right add_to_group_name
param it started failing on the contact id).

This just adjusts our code to the new normal

Bug: T231362

Change-Id: Ia7119b074940276884f9035bef3b8816d6669b63