Thank you test fixes
Contact ID is no longer settable in the form (we never used it).
Instead, we always send -1 to the mail send function, and use that
as an indication to not try to create a CiviMail record.

Also, a minor tweak to the 'ALL' locale, to only send in the langs
that exist for the selected template.

Bug: T232504
Change-Id: I8fdfac5591649b633f28de288169d9c940b27cc0