Add email amender extension
I have done quite a bit of cleanup and also added some apis to do this via the command line.

Note that through the UI I searched for domains ending in .co & found around 22k - it was slow but got there
as one would expect for a non-indexed query.

When I did the update in the UI it whitescreened but actually completed in the back ground. This
is arguably OK for a one-off thing but not a great look & the api should give us more tinker-power.

The unknown is how the api will perform with a complex regex. We know that using regex bypasses the index
but we are by definition bypassing the index here. We don't know how much additional degradation it creates.

We can iterate if it's too slow, or manage with the limit parameter.

Bug: T231332
Change-Id: Icf6ff9f5200e38bb2233ac28d84cc277c2cf62e2