EOY summary: split calculation, include more donations
The previous rollup only included one-time donations from contact_ids
that had a recurring donation. This one makes sure we include donations
for all contacts that share the email address, and include the newest
donor information across all contacts with that address, not just the
contact with the recurring donation.

This commit also fixes the job table so that the email column exactly
matches the definition of the main civi email column, and adds a
unique constraint to enable us to use ON DUPLICATE KEY IGNORE to
get the newest donor information.

This patch set uses a couple of temporary tables to optimize the
query speed. A previous patch set used subqueries and took 22
minutes on staging. This one takes about 4 mintues.

Bug: T195907
Change-Id: I75c06a9f9e6da437b0c0158137a3ca79ca88aa3d