Move logic for handling Yes-No fields during dedupe from wmf_civicrm to deduper.
This lays the groundwork for various conflicts we want to better handle by starting to set up the structure for handling them to the deduper
extension & make it more extendable.

As is it should result in no functional change but we should be able to easily add more fields to be resolved to 'yes' by
the script more easily.

A few thoughts

1) A few thoughts
1) I looked at how to align with the resolvers in https://github.com/systopia/de.systopia.xdedupe
- a fundamental difference, however, is that Bjoern's work conceptually involves resolving
conflicts and then calling the Contact.merge action. Our logic to date has been to rely on
'other things' (batch job, deduper, legacy form) calling the Merge action and then
to resolve conflicts via a hook. I don't see us switching as we currently have a scenario where we can
hook into Bjoern's screen or any other dedupe methodology, whereas his resolvers only work from his flow.
That makes it a little different to figure out how to integrate the two but I'm still pondering.

2) The resolution in the Deduper is site specific - ie ALWAYS resolve in this way in batch mode
(including the deduper but not the legacy form).

3) I need to tackle 'on_hold' for email - I suspect that is not resolving yet & potentially
add a test for that and the custom_data fields.

Bug: T235450

Change-Id: Ibf751f1c567c45eb14aea5ff1cb804bc79c569e4