Updated triggers for civicrm_value_1_prospect_5.mysql
Per https://phabricator.wikimedia.org/T236096 Nora deleted a column from this table & as the triggers were not updated
to match nothing can write to it now until they are updated. This is the updated trigger.

This should not require an outage as
1) it's only one table
2) it's already unwritable / broken & has been for about 12 hours at this point
3) none of the queues write to this table - hence it not burning down the house already.

I logged https://phabricator.wikimedia.org/T236111 to address the (low) risk of this happening again.

Bug: T236096
Change-Id: Ibbbe7c1455daff251a69753ee1f1de364db640ab