Add handling for resolving boolean conflict on 'email.on_hold
This adds generic boolean handling for email fields to the dedupetools extension.

Currently the setting ensures it only actually applies to on_hold for us and we have been
specifically NOT handling on_hold in the wmf_civicrm.module so with a little
early-exitting in there we can avoid the new preferrred place for resolving merge
issues from conflicting with the existing one which we aim to deprecate.

Bug: T221914

Change-Id: I9b1b67435537da496be17ec091246ad4c12f4b0c