Fix resolution of preferred contact language when no contributions present
We have been getting the odd failure on remind me later contacts not merging. Pretty sure this is the line causing it.
I did notice one that I manually merged had a language conflict & this is the only place using getsingle or getvalue with contribution,
other than a function that seems unused.

Bug: T235794
Change-Id: I63dfaace7bdbcb45b086c57e9fa6752dc54c7754