Introduce the concept of preferred contact, add ability to specify fields to be resolved that way.
The goal with this code is to resolve opt_in (and later some 'equivalent names'
to be resolved onto the 'preferred contact'. I added a few methods to choose that just
for general use but our preferred is most recent donor.

I have moved the handling of 'contact_source' over to this methodology.

However, I have commented out adding opt_in for now. There is some nasty stuff
the form layer does around yes-no fields & custom fields in terms of what we receive
from the hook so I'll tackle that in a follow up commit.

This is a bit of a big chunk but I think I added an OK explanation here
https://gerrit.wikimedia.org/r/#/c/wikimedia/fundraising/crm/+/547712/

Bug: T232037

Change-Id: I2694e320179fcd8a77df7f157890ba0c5a02997a