Add name_pairs table and resolvers
The name pairs table is a place to save pairs of names that are equivalent - these include
misspellings, nick names, script variations and ??

The intention is that the name_pair table will be saveable to from the Deduper (still thinking about the
appearance but if someone is deduping Bob & Bobby they should have a way to save them to the table).

This commit handles 'inferior' names and names where we resolve according to our preference.

 I originally conceptualised 'inferior'  as misspellings
but renamed the field in order to reflect the possibilty that a name might be expressed in
2 different scripts and both would be correct but we might prefer one (depending on the opinion of
the regional expert) or they might have equal weighting (and we would opt for the valuee from
the preferred contact).

If one name is not inferior the options for resolving them are as follows.

- prefer nick name, discard conficting name
- prefer non-nick name, discard conflicting name
- prefer non-nick name, put nick-name in nick name field
- prefer value from preferred contact (in our case most recent donor), discard conflicting value
- prefer value from preferred contact, put nick name, if exists in nick name field

I expect we would use the 4th one - but depending on stakeholders we might use the 5th one

Change-Id: Iabd2493fb58742cdd5dbd85de0ac1e74963fe067