Revisit Omnimail timestamp misery.
Erin has asked us to look into apparently missing data for some mailings in 2018. I did some digging on the
one with the largest discrepancy between sent & rows in the civicrm_mailing_provider_data table, which was
around 5 November 2018 and found that the rows after midday were missing :-( I recall we did have an issue with
that in the past & I thought we did a backfill but it seems to apply to this mailing & possibly others.

The backfill script had last run up to 24 April 2018 & Silverpop has data from 19 Aug 2018 still (it
gets deleted eventually). I haven't yet done a full audit but I did check Oct this year & it was all fine.

I tried re-importing some data around the earlies stuff Silverpop still has & pulled in around 30k rows on
28 Oct. However, when I checked on it I found another of our old friends ... the problem with the dates not
being interpretted quite the same & winding up with 2 rows. On further digging into this I found
the php date is being set differently to the mysql date. I have tweaked the repair script to allow for this handful of rows
to be repaired (it still only deletes rows with a fixed time gap but it permits greater manipulation
of the gap & the fix up based on the data I'm looking at.).

I also realised that 1) the easiest way to deal with the discrepancy is to force mysql & php to use the
UTC timezone when running the script. However, when doing backfill it's going to be easiest to match the
wrongness with last time's wrongness - so I added a parameter to manage that ...phew got all that out.

Bug: T236855
Change-Id: I29d7edd0598d268afb77f0f66f44b0dbb06e21cb